change upload process
when the user uploads a file, it will upload it to the temp bucket. once the user publishes the audio during the upload process, it will get moved into the public bucket. In the AWS S3 console, a lifecycle rule will state that the objects in the temp bucket will expire in a day.