rework db transaction
decided when creating an object, it will return the full view model by refetching from the database. the issue with the previous solution is that the id will not commit when it returns from the handler, because the transaction is committed after the pipeline.

The solution is the remove the db transaction pipeline, and handle the transaction manually in the handler.